[AM] separate CalculatorService into multiple services with separate concerns
